Seam
Searches all of a given image's pixels, both vertically and horizontally, in order to find a seam used for image resizing. A shortest path from one end of the picture to the other that has the least amount of energy, as defined by the dual gradient energy function, is determined and removed from the picture either horizontally or vertically.